Move PullRequestFinder into own file
