+++ gh_release_commenter.rb
@@ -1,28 +1,7 @@
 #!/usr/bin/env ruby
 
-require 'octokit'
 require './lib/merged_pull_request_finder'
-
-GITHUB_TOKEN = ENV['GITHUB_TOKEN']
-
-class PullRequestCommenter
-  def initialize(repo)
-    @client = Octokit::Client.new(:access_token => GITHUB_TOKEN)
-    @repo = repo
-  end
-
-  def add_comment_to_issues(issue_numbers, comment)
-    issue_numbers.each do |issue|
-      puts "Posting comment to ##{issue} in #{@repo}"
-      comment(issue, comment)
-    end
-  end
-
-  private
-  def comment(issue_number, content)
-    @client.add_comment(@repo, issue_number, content.to_s)
-  end
-end
+require './lib/pull_request_commenter'
 
 def usage
   <<-USAGE.gsub(/^ {4}/, '')
